Remove ServerEnvConfig, Move server.env tests into MavenTestLSPServerEnv, remove shared suite
